backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-cloud-bigquery
firebase-admin
pydantic